Move a helper method inside of constructor function
Set some boilerplate values (the client's api key and the appid) from
inside of the constructor function instead of using the helper.

Make use of the 'assign' module when there is some input needed, like
in the getUserListings() and getPrices() methods.

Note: I changed order of parameters in the test for 'getPrices()'
function on line 36, in order for it to pass.